Reorganiser ATDD-struktur: dsl/, features/ under atdd/
- Opprett atdd/dsl/ og flytt trygdeavtale.dsl.ts dit (parallelt med drivers/, steps/)
- Flytt features/ inn under atdd/features/ (alt ATDD samlet i én mappe)
- Sett featuresRoot eksplisitt så .features-gen/ får rene domenestier
- Oppdater imports, config, docs og unit tests
