+++ .github/workflows/e2e-tests.yml
@@ -33,7 +33,7 @@ on:
         type: boolean
         default: false
       run_bdd:
-        description: 'Run ONLY the opt-in ATDD/BDD example (features/ → --project=bdd) instead of chromium'
+        description: 'Run ONLY the opt-in ATDD/BDD example (atdd/features/ → --project=bdd) instead of chromium'
         required: false
         type: boolean
         default: false
