+++ playwright.config.ts
@@ -12,8 +12,8 @@ import { defineBddConfig } from 'playwright-bdd';
  *
  * Genererer Playwright-testfiler i .features-gen/ fra .feature-filene + step-
  * definisjonene. Dette er Lag 1 i Farleys fire-lags-modell:
- *   Lag 1: features/           (.feature-filer, strukturert tekst på norsk)
- *   Lag 2: atdd/steps + atdd/   (bindinger + DSL)
+ *   Lag 1: atdd/features/      (.feature-filer, strukturert tekst på norsk)
+ *   Lag 2: atdd/dsl + atdd/steps  (DSL + bindinger)
  *   Lag 3: atdd/drivers, pages/, helpers/  (protokolldrivere)
  *   Lag 4: docker-compose       (system under test)
  *
@@ -21,9 +21,14 @@ import { defineBddConfig } from 'playwright-bdd';
  * (kjøres av `npm run test:bdd`). `.features-gen/` er gitignored, så kjør alltid
  * `npm run test:bdd` (som kjører `npx bddgen` først) og aldri `playwright test
  * --project=bdd` direkte (kan ellers kjøre utdaterte/manglende genererte filer).
+ *
+ * featuresRoot settes eksplisitt slik at .features-gen/ speiler domene-mappene
+ * direkte (f.eks. .features-gen/trygdeavtale/...) uten et overflødig atdd/features/-
+ * prefiks.
  */
 const bddTestDir = defineBddConfig({
-  features: 'features/**/*.feature',
+  featuresRoot: 'atdd/features',
+  features: 'atdd/features/**/*.feature',
   steps: ['atdd/steps/*.ts', 'atdd/fixtures.ts'],
   language: 'no',
 });
